// src/drivers/enums/drivers.enum.ts
export enum DriverEvent {
  JOIN_ROUTE = 'join-route',
  JOINED_ROUTE = 'joined-route',
  LEAVE_ROUTE = 'leave-route',
  LEAVED_ROUTE = 'leaved-route',
  START_ROUTE = 'start-route',
  ROUTE_STARTED = 'route-started',
  LOCATION_UPDATE = 'location-update',
  BROADCAST_LOCATION = 'broadcast-location',
}